+++ src/components/repository/interview/RepositoryInterviewList.tsx
@@ -7,7 +7,7 @@ import useGetAccessToken from '@/hooks/useGetAccessToken';
 import useCheckAccess from '@/hooks/useCheckExistAccess';
 import CheckDeleteInterviewModal from '../checkDeleteModal/checkDeleteInterviewModal/CheckDeleteInterviewModal';
 import { parseISO, format } from 'date-fns';
-import { useState } from 'react';
+import { useRepositoryInterviewList } from '@/hooks/repository/interview/useRepositoryInterviewList';
 import './RepositoryInterviewList.scss';
 
 interface InterviewResponse {
@@ -19,11 +19,9 @@ interface InterviewItem {
   level: 'EASY' | 'NORMAL' | 'HARD';
   createdAt: string;
 }
-
 interface InterviewDetailResponse {
   data: InterviewDetail;
 }
-
 interface InterviewDetail {
   id: number;
   title: string;
@@ -31,7 +29,6 @@ interface InterviewDetail {
   createdAt: string;
   questions: InterviewQuestion[];
 }
-
 interface InterviewQuestion {
   questionId: number;
   question: string;
@@ -43,23 +40,26 @@ const RepositoryInterviewList = () => {
   const accessToken = useGetAccessToken();
   const existAccess = useCheckAccess(accessToken);
   const { interviewDate } = useRepositoryDateStore();
-  const [expandedInterviewId, setExpandedInterviewId] = useState<number | null>(null);
-  const [visibleAnswerMap, setVisibleAnswerMap] = useState<Record<number, boolean>>({});
-  const [selectedInterviewIds, setSelectedInterviewIds] = useState<number[]>([]);
-  const [shakeDelete, setShakeDelete] = useState(false);
-  const [showDeleteModal, setShowDeleteModal] = useState(false);
 
-
-  const today = new Date();
-  const yyyy = today.getFullYear();
-  const mm = String(today.getMonth() + 1).padStart(2, '0');
-  const dd = String(today.getDate()).padStart(2, '0');
-  const formattedToday = `${yyyy}-${mm}-${dd}`;
+  const {
+    expandedInterviewId,
+    visibleAnswerMap,
+    setVisibleAnswerMap,
+    selectedInterviewIds,
+    setSelectedInterviewIds,
+    shakeDelete,
+    showDeleteModal,
+    setShowDeleteModal,
+    mapLevelToLabel,
+    handleClickInterview,
+    toggleInterviewSelection,
+    handleDeleteClick,
+  } = useRepositoryInterviewList();
 
   const { data: interviewData } = useQuery<InterviewItem[]>({
     queryKey: ['interview-list', interviewDate],
     queryFn: async () => {
-      const targetDate = interviewDate || formattedToday;
+      const targetDate = interviewDate || format(new Date(), 'yyyy-MM-dd');
       const response = await callApi<InterviewResponse>({
         method: 'GET',
         endpoint: `/interviews?page=0&size=10&date=${targetDate}`,
@@ -90,55 +90,20 @@ const RepositoryInterviewList = () => {
     gcTime: 3600000,
   });
 
-  const mapLevelToLabel = (level: 'EASY' | 'NORMAL' | 'HARD'): string => {
-    switch (level) {
-      case 'EASY': return '쉬움';
-      case 'NORMAL': return '보통';
-      case 'HARD': return '어려움';
-      default: return '-';
-    }
-  };
-
-  const handleClickInterview = (interviewId: number) => {
-    if (selectedInterviewIds.includes(interviewId)) return;
-    setExpandedInterviewId(prev => (prev === interviewId ? null : interviewId));
-  };
-
-  const toggleInterviewSelection = (interviewId: number) => {
-    setSelectedInterviewIds(prev =>
-      prev.includes(interviewId)
-        ? prev.filter(id => id !== interviewId)
-        : [...prev, interviewId]
-    );
-    if (expandedInterviewId === interviewId) {
-      setExpandedInterviewId(null);
-    }
-  };
-
-  const handleDeleteClick = () => {
-    if (selectedInterviewIds.length === 0) {
-      setShakeDelete(true);
-      setTimeout(() => setShakeDelete(false), 500);
-    } else {
-      setShowDeleteModal(true);
-    }
-  };
   return (
     <div className="repository-interview-list">
       <div className="repository-interview-list__header">
         <h2 className="repository-interview-list__title">면접 질문 목록</h2>
-        {interviewData && interviewData.length > 0 && (
+        {!!interviewData?.length && (
           <button
             className={`repository-interview-list__button${shakeDelete ? ' error shake' : ''}`}
             onClick={handleDeleteClick}
-            style={{ visibility: interviewData && interviewData.length > 0 ? 'visible' : 'hidden' }}
           >
             삭제
           </button>
         )}
       </div>
       <ul className="repository-interview-list__items">
-
         {interviewData?.map((interview) => {
           const formattedDate = format(parseISO(interview.createdAt), 'yyyy-MM-dd : HH:mm:ss');
           const isExpanded = expandedInterviewId === interview.id;
@@ -155,13 +120,12 @@ const RepositoryInterviewList = () => {
                   onClick={() => handleClickInterview(interview.id)}
                 >
                   <div className="repository-interview-list__item-header-top">
-                  <h3 className={`repository-interview-list__item-title${ isExpanded ? ' repository-interview-list__item-title--expanded' : ''}`}>
-                    [{mapLevelToLabel(interview.level)}] {interview.title}
-                  </h3>
+                    <h3 className={`repository-interview-list__item-title${isExpanded ? ' repository-interview-list__item-title--expanded' : ''}`}>
+                      [{mapLevelToLabel(interview.level)}] {interview.title}
+                    </h3>
                   </div>
                   <p className="repository-interview-list__item-date">{formattedDate}</p>
                 </div>
-
                 <input
                   type="checkbox"
                   className="repository-interview-list__item-checkbox"
@@ -177,7 +141,7 @@ const RepositoryInterviewList = () => {
                       <p className="repository-interview-list__item-question">{q.question}</p>
                       <button
                         onClick={() =>
-                          setVisibleAnswerMap(prev => ({
+                          setVisibleAnswerMap((prev) => ({
                             ...prev,
                             [q.questionId]: !prev[q.questionId],
                           }))
@@ -197,6 +161,7 @@ const RepositoryInterviewList = () => {
           );
         })}
       </ul>
+
       {showDeleteModal && (
         <CheckDeleteInterviewModal
           interviewIds={selectedInterviewIds}
