+++ src/hooks/repository/interview/useRepositoryInterviewList.tsx
@@ -1,0 +1,60 @@
+'use client';
+
+import { useState } from 'react';
+
+export const useRepositoryInterviewList = () => {
+  const [expandedInterviewId, setExpandedInterviewId] = useState<number | null>(null);
+  const [visibleAnswerMap, setVisibleAnswerMap] = useState<Record<number, boolean>>({});
+  const [selectedInterviewIds, setSelectedInterviewIds] = useState<number[]>([]);
+  const [shakeDelete, setShakeDelete] = useState(false);
+  const [showDeleteModal, setShowDeleteModal] = useState(false);
+
+  const mapLevelToLabel = (level: 'EASY' | 'NORMAL' | 'HARD'): string => {
+    switch (level) {
+      case 'EASY': return '쉬움';
+      case 'NORMAL': return '보통';
+      case 'HARD': return '어려움';
+      default: return '-';
+    }
+  };
+
+  const handleClickInterview = (interviewId: number) => {
+    if (selectedInterviewIds.includes(interviewId)) return;
+    setExpandedInterviewId(prev => (prev === interviewId ? null : interviewId));
+  };
+
+  const toggleInterviewSelection = (interviewId: number) => {
+    setSelectedInterviewIds(prev =>
+      prev.includes(interviewId)
+        ? prev.filter(id => id !== interviewId)
+        : [...prev, interviewId]
+    );
+    if (expandedInterviewId === interviewId) {
+      setExpandedInterviewId(null);
+    }
+  };
+
+  const handleDeleteClick = () => {
+    if (selectedInterviewIds.length === 0) {
+      setShakeDelete(true);
+      setTimeout(() => setShakeDelete(false), 500);
+    } else {
+      setShowDeleteModal(true);
+    }
+  };
+
+  return {
+    expandedInterviewId,
+    visibleAnswerMap,
+    setVisibleAnswerMap,
+    selectedInterviewIds,
+    setSelectedInterviewIds,
+    shakeDelete,
+    showDeleteModal,
+    setShowDeleteModal,
+    mapLevelToLabel,
+    handleClickInterview,
+    toggleInterviewSelection,
+    handleDeleteClick,
+  };
+};
